added silver_vote.py
